+++ www/wp-content/themes/luziapi/src/Pilotage/Bootstrap/PilotageServiceProvider.php
@@ -18,7 +18,7 @@ use LuziApi\Pilotage\Application\Command\RecordOrderStockMovement\OrderStockMove
 use LuziApi\Pilotage\Application\Command\RecordReceipt\RecordReceiptHandler;
 use LuziApi\Pilotage\Application\Command\RecordStockMovement\RecordStockMovementHandler;
 use LuziApi\Pilotage\Application\Command\ReverseReceipt\ReverseReceiptHandler;
-use LuziApi\Pilotage\Application\Command\UpdateCustomerContact\UpdateCustomerContactHandler;
+use LuziApi\Pilotage\Application\Command\SaveCustomerProfile\SaveCustomerProfileHandler;
 use LuziApi\Pilotage\Application\Query\GetActivityLog\GetActivityLogHandler;
 use LuziApi\Pilotage\Application\Query\GetAnnualDashboard\GetAnnualDashboardHandler;
 use LuziApi\Pilotage\Application\Query\GetCustomerDirectory\GetCustomerDirectoryHandler;
@@ -35,7 +35,6 @@ use LuziApi\Pilotage\Domain\Receipt\ReceiptReconciliationProjector;
 use LuziApi\Pilotage\Domain\Sales\AnnualSalesCalculator;
 use LuziApi\Pilotage\Domain\Tax\MicroBaCalculator;
 use LuziApi\Pilotage\Infrastructure\WooCommerce\WooCommerceActivitySubscriber;
-use LuziApi\Pilotage\Infrastructure\WooCommerce\WooCommerceCustomerContactWriter;
 use LuziApi\Pilotage\Infrastructure\WooCommerce\WooCommerceCustomerTimelineRepository;
 use LuziApi\Pilotage\Infrastructure\WooCommerce\WooCommerceLoyaltyEconomicsReader;
 use LuziApi\Pilotage\Infrastructure\WooCommerce\WooCommerceOrderDiscountWriter;
@@ -54,6 +53,7 @@ use LuziApi\Pilotage\Infrastructure\WordPress\PilotageSchemaManager;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressActivityRepository;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressClock;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressCustomerCategoryRepository;
+use LuziApi\Pilotage\Infrastructure\WordPress\WordPressCustomerProfileRepository;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressInventoryRepository;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressReceiptRepository;
 use LuziApi\Pilotage\Infrastructure\WordPress\WordPressTaxSettings;
@@ -103,6 +103,7 @@ final class PilotageServiceProvider
             new WordPressCustomerCategoryRepository($wpdb, $schema),
             $activity,
         );
+        $customerProfiles = new WordPressCustomerProfileRepository($wpdb, $schema);
         $receipts = new AuditedReceiptRepository(
             new WordPressReceiptRepository($wpdb, $schema, $clock->timezone()),
             $activity,
@@ -130,6 +131,7 @@ final class PilotageServiceProvider
             $receipts,
             new WooCommerceCustomerTimelineRepository($clock->timezone()),
             $clock,
+            $customerProfiles,
         );
         $receiptHandler = new GetReceiptRegisterHandler(
             $receipts,
@@ -199,7 +201,7 @@ final class PilotageServiceProvider
             ),
             \LuziApi\Loyalty\Bootstrap\LoyaltyServiceProvider::loyaltyAdjustmentHandler(),
             $subscribers,
-            new UpdateCustomerContactHandler(new WooCommerceCustomerContactWriter()),
+            new SaveCustomerProfileHandler($customerProfiles, $clock),
         );
         $loyaltyController = new LoyaltyController(new GetLoyaltyDashboardHandler(
             $orders,